+++ packages/ranges-is-index-within/rollup.config.js
@@ -44,14 +44,7 @@ export default commandLineArgs => {
     {
       input: "src/main.js",
       output: [{ file: pkg.main, format: "cjs" }],
-      external: [
-        "check-types-mini",
-        "is-natural-number",
-        "is-natural-number-string",
-        "lodash.isplainobject",
-        "ordinal-number-suffix",
-        "ranges-sort"
-      ],
+      external: [""],
       plugins: [
         strip({
           sourceMap: false
@@ -69,14 +62,7 @@ export default commandLineArgs => {
     {
       input: "src/main.js",
       output: [{ file: pkg.module, format: "es" }],
-      external: [
-        "check-types-mini",
-        "is-natural-number",
-        "is-natural-number-string",
-        "lodash.isplainobject",
-        "ordinal-number-suffix",
-        "ranges-sort"
-      ],
+      external: [""],
       plugins: [
         strip({
           sourceMap: false